+++ public/js/user.structure.js
@@ -21,7 +21,7 @@ const userboxUI = {
   icon(iconNameWithoutEva){
     return new TagString(`<i class="eva eva-${iconNameWithoutEva}"></i>`)
   },
-  input(id=''){
-    return new TagString(`<input type="checkbox" id="${id}">`);
+  input(id='', isRadio = false){
+    return new TagString(`<input type="${isRadio ? 'radio' : 'checkbox'}" id="${id}" />`);
   }
 }